backend/requirements.txt
fastapi
uvicorn
scikit-learn
web3
python-dotenv
